# docs/api-specs/graph-service.openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Graph Engine API",
        "description": "Address clustering and tag propagation service for Chain Risk Platform",
        "contact": {
            "name": "Chain Risk Team",
            "email": "chainrisk@example.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Default Server"
        }
    ],
...
    "paths": {
...
        "/api/graph/propagate/{address}": {
            "post": {
                "tags": [
                    "Graph API"
                ],
                "summary": "Propagate from address",
                "description": "Propagate risk from a specific source address",
                "operationId": "propagateFromAddress",
                "parameters": [
                    {
                        "name": "address",
                        "in": "path",
                        "description": "Source address",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PropagationResultResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/graph/cluster/manual": {
            "post": {
                "tags": [
                    "Graph API"
                ],
                "summary": "Manually cluster addresses",
                "description": "Create a cluster from a list of addresses",
                "operationId": "manualCluster",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClusteringResultResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/graph/address/{address}/tags": {
            "get": {
                "tags": [
                    "Graph API"
                ],
                "summary": "Get address tags",
                "description": "Get all tags associated with an address",
                "operationId": "getAddressTags",
                "parameters": [
                    {
                        "name": "address",
                        "in": "path",
                        "description": "Blockchain address",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Graph API"
                ],
                "summary": "Add tags to address",
                "description": "Add one or more tags to an address (manual tagging)",
                "operationId": "addTags",
                "parameters": [
                    {
                        "name": "address",
                        "in": "path",
                        "description": "Blockchain address",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddTagRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AddressInfoResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/graph/search/tag/{tag}": {
            "get": {
                "tags": [
                    "Graph API"
                ],
                "summary": "Search by tag",
                "description": "Find addresses with a specific tag",
                "operationId": "searchByTag",
                "parameters": [
                    {
                        "name": "tag",
                        "in": "path",
                        "description": "Tag to search for",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Maximum results",
                        "required": false,
                        "schema": {
                            "maximum": 200,
                            "minimum": 1,
                            "type": "integer",
                            "format": "int32",
                            "default": 50
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AddressInfoResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/graph/search/high-risk": {
            "get": {
                "tags": [
                    "Graph API"
                ],
                "summary": "Get high-risk addresses",
                "description": "Find addresses with risk score above threshold",
                "operationId": "getHighRiskAddresses",
                "parameters": [
                    {
                        "name": "threshold",
                        "in": "query",
                        "description": "Risk score threshold (0.0 - 1.0)",
                        "required": false,
                        "schema": {
                            "maximum": 1,
                            "minimum": 0,
                            "type": "number",
                            "format": "double",
                            "default": 0.6
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Maximum results",
                        "required": false,
                        "schema": {
                            "maximum": 200,
                            "minimum": 1,
                            "type": "integer",
                            "format": "int32",
                            "default": 50
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AddressInfoResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/graph/path/{fromAddress}/{toAddress}": {
            "get": {
                "tags": [
                    "Graph API"
                ],
                "summary": "Find path between addresses",
                "description": "Find the shortest path of transfers between two addresses",
                "operationId": "findPath",
                "parameters": [
                    {
                        "name": "fromAddress",
                        "in": "path",
                        "description": "Source address",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "toAddress",
                        "in": "path",
                        "description": "Target address",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "maxDepth",
                        "in": "query",
                        "description": "Maximum path depth",
                        "required": false,
                        "schema": {
                            "maximum": 10,
                            "minimum": 1,
                            "type": "integer",
                            "format": "int32",
                            "default": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PathResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/graph/cluster/{clusterId}": {
            "get": {
                "tags": [
                    "Graph API"
                ],
                "summary": "Get cluster by ID",
                "description": "Get cluster information by cluster ID",
                "operationId": "getCluster",
                "parameters": [
                    {
                        "name": "clusterId",
                        "in": "path",
                        "description": "Cluster ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Cluster not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClusterResponse"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Cluster found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClusterResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/graph/address/{address}": {
            "get": {
                "tags": [
                    "Graph API"
                ],
                "summary": "Get address information",
                "description": "Get detailed information about an address including tags and risk score",
                "operationId": "getAddressInfo",
                "parameters": [
                    {
                        "name": "address",
                        "in": "path",
                        "description": "Blockchain address (0x-prefixed)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Address found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AddressInfoResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Address not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AddressInfoResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/graph/address/{address}/neighbors": {
            "get": {
                "tags": [
                    "Graph API"
                ],
                "summary": "Get address neighbors",
                "description": "Get addresses that have transferred to/from this address",
                "operationId": "getNeighbors",
                "parameters": [
                    {
                        "name": "address",
                        "in": "path",
                        "description": "Blockchain address",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "depth",
                        "in": "query",
                        "description": "Search depth (1-3)",
                        "required": false,
                        "schema": {
                            "maximum": 3,
                            "minimum": 1,
                            "type": "integer",
                            "format": "int32",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Maximum number of neighbors to return",
                        "required": false,
                        "schema": {
                            "maximum": 200,
                            "minimum": 1,
                            "type": "integer",
                            "format": "int32",
                            "default": 50
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AddressNeighborsResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/graph/address/{address}/cluster": {
            "get": {
                "tags": [
                    "Graph API"
                ],
                "summary": "Get address cluster",
                "description": "Get the cluster this address belongs to",
                "operationId": "getAddressCluster",
                "parameters": [
                    {
                        "name": "address",
                        "in": "path",
                        "description": "Blockchain address",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Address not in any cluster",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClusterResponse"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Cluster found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClusterResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/graph/address/{address}/tags/{tag}": {
            "delete": {
                "tags": [
                    "Graph API"
                ],
                "summary": "Remove tag from address",
                "description": "Remove a specific tag from an address",
                "operationId": "removeTag",
                "parameters": [
                    {
                        "name": "address",
                        "in": "path",
                        "description": "Blockchain address",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tag",
                        "in": "path",
                        "description": "Tag to remove",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "PropagationResultResponse": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string"
                    },
                    "addressesAffected": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "tagsPropagated": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "maxHops": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "decayFactor": {
                        "type": "number",
                        "format": "double"
                    },
                    "durationMs": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "startedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "completedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "errorMessage": {
                        "type": "string"
                    }
                }
            },
            "ClusteringResultResponse": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string"
                    },
                    "clustersCreated": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "addressesClustered": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "durationMs": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "startedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "completedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "errorMessage": {
                        "type": "string"
                    }
                }
            },
            "AddTagRequest": {
                "required": [
                    "tags"
                ],
                "type": "object",
                "properties": {
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "source": {
                        "type": "string"
                    },
                    "confidence": {
                        "type": "number",
                        "format": "double"
                    }
                }
            },
            "AddressInfoResponse": {
                "type": "object",
                "properties": {
                    "address": {
                        "type": "string"
                    },
                    "firstSeen": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "lastSeen": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "txCount": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "riskScore": {
                        "type": "number",
                        "format": "double"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "clusterId": {
                        "type": "string"
                    },
                    "network": {
                        "type": "string"
                    },
                    "incomingCount": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "outgoingCount": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
            "PathResponse": {
                "type": "object",
                "properties": {
                    "found": {
                        "type": "boolean"
                    },
                    "fromAddress": {
                        "type": "string"
                    },
                    "toAddress": {
                        "type": "string"
                    },
                    "pathLength": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "maxDepth": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "path": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PathNode"
                        }
                    }
                }
            },
...
        }
    }
}